project-rag/requirements.txt
python-dotenv
streamlit
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
pymupdf
groq
sentence-transformer